refactor: menu link component
